Refactor rendering
