+++ rest-api-spec/src/main/resources/rest-api-spec/api/search_application.render_metadata.json
@@ -1,0 +1,38 @@
+{
+  "search_application.render_metadata": {
+    "documentation": {
+      "url": "https://www.elastic.co/guide/en/elasticsearch/reference/master/search-application-render-metadata.html",
+      "description": "Renders the metadata for a given search query against the default template parameters"
+    },
+    "stability": "beta",
+    "visibility": "public",
+    "headers": {
+      "accept": [
+        "application/json"
+      ],
+      "content_type": [
+        "application/json"
+      ]
+    },
+    "url": {
+      "paths": [
+        {
+          "path": "/_application/search_application/{name}/_render_metadata",
+          "methods": [
+            "POST"
+          ],
+          "parts": {
+            "name": {
+              "type": "string",
+              "description": "The name of the search application to render metadata for"
+            }
+          }
+        }
+      ]
+    },
+    "body": {
+      "description": "Search parameters, including template parameters that override defaults",
+      "required": false
+    }
+  }
+}
