+++ x-pack/plugin/ent-search/qa/rest/src/yamlRestTest/resources/rest-api-spec/test/entsearch/52_search_application_render_metadata.yml
@@ -1,0 +1,148 @@
+setup:
+  - do:
+      indices.create:
+        index: test-search-index1
+        body:
+          settings:
+            index:
+              number_of_shards: 1
+              number_of_replicas: 0
+
+  - do:
+      indices.create:
+        index: test-search-index2
+        body:
+          settings:
+            index:
+              number_of_shards: 1
+              number_of_replicas: 0
+
+  - do:
+      indices.create:
+        index: test-index
+        body:
+          settings:
+            index:
+              number_of_shards: 1
+              number_of_replicas: 0
+
+  - do:
+      search_application.put:
+        name: test-search-application
+        body:
+          indices: [ "test-search-index1", "test-search-index2" ]
+          analytics_collection_name: "test-analytics"
+          template:
+            script:
+              source:
+                query:
+                  term:
+                    "{{field_name}}": "{{field_value}}"
+              params:
+                field_name: field1
+                field_value: value1
+            dictionary:
+              field_name:
+                type: string
+              field_value:
+                type: string
+
+  - do:
+      index:
+        index: test-search-index1
+        id: doc1
+        body:
+          field1: value1
+          field2: value1
+        refresh: true
+
+  - do:
+      index:
+        index: test-search-index2
+        id: doc2
+        body:
+          field1: value1
+          field3: value3
+        refresh: true
+
+---
+teardown:
+  - do:
+      search_application.delete:
+        name: test-search-application
+        ignore: 404
+
+  - do:
+      indices.delete:
+        index: test-search-index1
+        ignore: 404
+
+  - do:
+      indices.delete:
+        index: test-search-index2
+        ignore: 404
+
+  - do:
+      indices.delete:
+        index: test-index
+        ignore: 404
+
+---
+"Render Metadata for Search Application with default parameters":
+
+  - do:
+      search_application.render_metadata:
+        name: test-search-application
+
+  - match: { query.term.field1: "value1" }
+
+---
+"Render Metadata for Search Application overriding part of the parameters":
+
+  - do:
+      search_application.render_metadata:
+        name: test-search-application
+        body:
+          params:
+            field_name: field2
+
+
+  - match: { query.term.field2: "value1" }
+
+---
+"Render Metadata for Search Application overriding all parameters":
+
+  - do:
+      search_application.render_metadata:
+        name: test-search-application
+        body:
+          params:
+            field_name: field3
+            field_value: value3
+
+
+  - match: { query.term.field3: "value3" }
+
+---
+"Render Metadata for Search Application with invalid parameter validation":
+
+  - do:
+      catch: "bad_request"
+      search_application.render_metadata:
+        name: test-search-application
+        body:
+          params:
+            field_name: field3
+            field_value: 35
+
+---
+"Render Metadata for Search Application - not found":
+
+  - do:
+      catch: "missing"
+      search_application.render_metadata:
+        name: nonexisting-test-search-application
+        body:
+          params:
+            field_name: field3
+            field_value: value3
