+++ x-pack/plugin/ent-search/src/main/java/org/elasticsearch/xpack/application/search/action/RestQuerySearchApplicationAction.java
@@ -12,7 +12,6 @@ import org.elasticsearch.rest.BaseRestHandler;
 import org.elasticsearch.rest.RestRequest;
 import org.elasticsearch.rest.action.RestToXContentListener;
 import org.elasticsearch.xpack.application.EnterpriseSearch;
-import org.elasticsearch.xpack.application.search.action.QuerySearchApplicationAction.Request;
 
 import java.io.IOException;
 import java.util.List;
@@ -37,13 +36,13 @@ public class RestQuerySearchApplicationAction extends BaseRestHandler {
     @Override
     protected RestChannelConsumer prepareRequest(RestRequest restRequest, NodeClient client) throws IOException {
         final String searchAppName = restRequest.param("name");
-        Request request;
+        SearchApplicationSearchRequest request;
         if (restRequest.hasContent()) {
-            request = Request.fromXContent(searchAppName, restRequest.contentParser());
+            request = SearchApplicationSearchRequest.fromXContent(searchAppName, restRequest.contentParser());
         } else {
-            request = new Request(searchAppName);
+            request = new SearchApplicationSearchRequest(searchAppName);
         }
-        final Request finalRequest = request;
+        final SearchApplicationSearchRequest finalRequest = request;
         return channel -> client.execute(QuerySearchApplicationAction.INSTANCE, finalRequest, new RestToXContentListener<>(channel));
     }
 }
