+++ x-pack/plugin/ent-search/src/main/java/org/elasticsearch/xpack/application/search/action/RestRenderMetadataAction.java
@@ -35,7 +35,10 @@ public class RestRenderMetadataAction extends BaseRestHandler {
     @Override
     protected RestChannelConsumer prepareRequest(RestRequest restRequest, NodeClient client) throws IOException {
         final String searchAppName = restRequest.param("name");
-        final RenderMetadataAction.Request request = RenderMetadataAction.Request.fromXContent(searchAppName, restRequest.contentParser());
+        final SearchApplicationSearchRequest request = SearchApplicationSearchRequest.fromXContent(
+            searchAppName,
+            restRequest.contentParser()
+        );
         return channel -> client.execute(RenderMetadataAction.INSTANCE, request, new RestToXContentListener<>(channel));
     }
 }
